# fix: cache const response