--- a/src/core/app_context.rs
+++ b/src/core/app_context.rs
@@ -31,6 +31,7 @@ pub struct AppContext {
     pub dedupe_handler: Arc<DedupeResult<IoId, ConstValue, Error>>,
     pub dedupe_operation_handler: DedupeResult<OperationId, AnyResponse<Vec<u8>>, Error>,
     pub operation_plans: DashMap<OPHash, OperationPlan<async_graphql_value::Value>>,
+    pub const_execution_cache: DashMap<OPHash, AnyResponse<Vec<u8>>>,
 }
 
 impl AppContext {
@@ -152,6 +153,7 @@ impl AppContext {
             dedupe_handler: Arc::new(DedupeResult::new(false)),
             dedupe_operation_handler: DedupeResult::new(false),
             operation_plans: DashMap::new(),
+            const_execution_cache: DashMap::default(),
         }
     }