--- a/src/core/jit/exec_const.rs
+++ b/src/core/jit/exec_const.rs
@@ -20,13 +20,12 @@ use crate::core::Transform;
 
 /// A specialized executor that executes with async_graphql::Value
 pub struct ConstValueExecutor {
-    pub plan: OperationPlan<Value>,
-    pub response: Option<Response<ConstValue>>,
+    plan: OperationPlan<Value>,
 }
 
 impl From<OperationPlan<Value>> for ConstValueExecutor {
     fn from(plan: OperationPlan<Value>) -> Self {
-        Self { plan, response: None }
+        Self { plan }
     }
 }
 
@@ -37,12 +36,11 @@ impl ConstValueExecutor {
     }
 
     pub async fn execute(
-        mut self,
+        self,
         req_ctx: &RequestContext,
         request: &Request<ConstValue>,
     ) -> Response<ConstValue> {
         let variables = &request.variables;
-        let is_const = self.plan.is_const;
 
         // Attempt to skip unnecessary fields
         let Ok(plan) = transform::Skip::new(variables)
@@ -77,11 +75,6 @@ impl ConstValueExecutor {
         let store = exe.store().await;
         let synth = Synth::new(&plan, store, vars);
         let response = exe.execute(synth).await;
-
-        // Cache the response if we know the output is always the same
-        if is_const {
-            self.response = Some(response.clone());
-        }
 
         response
     }