--- a/src/core/jit/exec_const.rs
+++ b/src/core/jit/exec_const.rs
@@ -74,9 +74,9 @@ impl ConstValueExecutor {
         let exe = Executor::new(&plan, exec);
         let store = exe.store().await;
         let synth = Synth::new(&plan, store, vars);
-        let response = exe.execute(synth).await;
+        
 
-        response
+        exe.execute(synth).await
     }
 }