@laststylebender14 2024-10-24
## Before
```lua
➜  virtual-federation git:(main) ✗ wrk -t4 -c100 -d10s -s bench-wrk.lua http://127.0.0.1:8000/graphql
Running 10s test @ http://127.0.0.1:8000/graphql
  4 threads and 100 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     1.94ms    7.44ms 151.03ms   94.93%
    Req/Sec    63.84k    36.80k  282.05k    70.77%
  2530928 requests in 10.07s, 340.33MB read
Requests/sec: 251415.96
Transfer/sec:     33.81MB
```

## After

```lua
➜  virtual-federation git:(main) ✗ wrk -t4 -c100 -d10s -s bench-wrk.lua http://127.0.0.1:8000/graphql
Running 10s test @ http://127.0.0.1:8000/graphql
  4 threads and 100 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   304.60us  111.33us   8.27ms   76.81%
    Req/Sec    70.75k     3.47k   79.98k    80.45%
  2844202 requests in 10.10s, 382.45MB read
Requests/sec: 281574.44
Transfer/sec:     37.86MB
```